requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.1
streamlit==1.40.0
pandas==2.2.3
altair==5.4.1

